update Greedy Dispense Strategy

+++ src/main/java/com.atm/strategy/DispenseStrategy.java
@@ -3,5 +3,5 @@ package com.atm.strategy;
 import java.util.Map;
 
 public interface DispenseStrategy {
-    Map<Integer, Integer> dispense(double amount);
+    Map<Integer, Integer> dispense(int amount);
 }
